Refactor Cloudinary module structure by updating imports and removing unused files
- Changed import paths for UploadedFileType and CloudinaryImageService to reflect the new directory structure under 'infrastructure/services'.
- Removed obsolete CloudinaryImageService and CloudinaryService files, streamlining the module and enhancing maintainability.
- Updated related test files to ensure they reference the correct imports, improving overall code clarity.

These changes improve the organization of the Cloudinary module and facilitate easier future updates.
