+++ src/modules/cloudinary/domain/interfaces/cache-entry.interface.ts
