+++ src/modules/cloudinary/domain/interfaces/file-upload.interface.ts
