+++ src/modules/cloudinary/domain/interfaces/folder-response.interface.ts
