+++ src/modules/cloudinary/domain/interfaces/resource-response.interface.ts
