+++ src/modules/cloudinary/infrastructure/services/cloudinary-image.service.ts
